[refactor/PPFR-1]: changed structure and imports
