--- src/pages/Authorized/index.ts
@@ -1,1 +1,0 @@
-export { AuthorizedPage } from './ui/AuthorizedPage';
