--- src/pages/Welcome/index.ts
@@ -1,1 +1,0 @@
-export { WelcomePage } from './ui/WelcomePage';
